refactor: move yaml common tag initialization stuff to new value_util.c gh-28
