+++ src/types/asdf_common_tag_map.h
@@ -5,7 +5,7 @@
 
 #include <stc/cstr.h>
 
-#include "../value.h"
+#include "../value_util.h"
 
 #define i_type asdf_common_tag_map
 #define i_keypro cstr
