+++ tests/test-value.c
@@ -9,18 +9,6 @@
 #include "util.h"
 
 
-/* TODO: Move this function and the test for it to some value_util module */
-/*
-MU_TEST(test_asdf_common_tag_get) {
-    asdf_yaml_common_tag_t tag = asdf_common_tag_get("tag:yaml.org,2002:str");
-    assert_int(tag, ==, ASDF_YAML_COMMON_TAG_STR);
-    asdf_yaml_common_tag_t unknown = asdf_common_tag_get("unknown tag");
-    assert_int(unknown, ==, ASDF_YAML_COMMON_TAG_UNKNOWN);
-    return MUNIT_OK;
-}
-*/
-
-
 MU_TEST(test_asdf_value_as_int64) {
     const char *path = get_fixture_file_path("scalars.asdf");
     asdf_file_t *file = asdf_open_file(path, "r");
@@ -40,7 +28,6 @@ MU_TEST(test_asdf_value_as_int64) {
 
 MU_TEST_SUITE(
     test_asdf_value,
-    //MU_RUN_TEST(test_asdf_common_tag_get),
     MU_RUN_TEST(test_asdf_value_as_int64)
 );
